requirements.txt
jsonrpc_requests
chainside-btcpy
git+https://github.com/trezor/python-trezor.git
web3
py-solc

